% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESizeCoHedg.R
\name{ESizeCoHedg}
\alias{ESizeCoHedg}
\title{Calculating the Effect Size of Cohen's d & Hedges' g}
\usage{
ESizeCoHedg()
}
\arguments{
\item{x1}{omitted scores of group1}

\item{x2}{omitted scores of group2}
}
\value{
Effect Size of Cohen's d & Hedges' g
}
\description{
\code{ESizeCoHedg} calculate the Effect Size & draw the dnorm curve
}
\examples{
# ESizeCoHedg(X1, X2)
}
